refactor: extract shared FilmPosterCard component
Move poster card layout, ScreeningTime, and DayRow into a shared
FilmPosterCard component used by both MovieCard and ByCinemaView.
Accepts a className prop to allow per-context background colors
(neutral-800 in movie view, neutral-700 inside cinema sections).
