+++ src/components/ByCinemaView.tsx
@@ -1,6 +1,5 @@
 import { MapPin } from 'lucide-react';
 import type { CinemaTable, FilmWithPoster, ShowingsTable } from '../api';
-import { formatDate, formatTime } from '../utils/formatters';
 import {
     groupByCinema,
     groupByDay,
@@ -8,23 +7,12 @@ import {
     sortGroupedByStartTime,
     sortScreeningByStartTime,
 } from '../utils/grouping';
+import FilmPosterCard, { DayRow } from './FilmPosterCard';
 
 interface ByCinemaViewProps {
     screenings: ShowingsTable[];
     getMovie: (id: number) => FilmWithPoster | undefined;
     getCinema: (id: number) => CinemaTable | undefined;
-}
-
-function ScreeningTime({ screening }: { screening: ShowingsTable }) {
-    const time = formatTime(screening.start_time);
-    const cls = 'text-red-500 inline';
-    return screening.booking_url ? (
-        <a className={`${cls} underline`} href={screening.booking_url} target="_blank">
-            {time}
-        </a>
-    ) : (
-        <p className={cls}>{time}</p>
-    );
 }
 
 export default function ByCinemaView({ screenings, getMovie, getCinema }: ByCinemaViewProps) {
@@ -49,56 +37,16 @@ export default function ByCinemaView({ screenings, getMovie, getCinema }: ByCine
                                     groupByDay(movieScreenings.sort(sortScreeningByStartTime))
                                 ).sort(sortGroupedByStartTime);
                                 return (
-                                    <div
-                                        key={key}
-                                        className="bg-neutral-600 rounded-lg overflow-hidden flex flex-col w-30 sm:w-40"
-                                    >
-                                        {movie?.poster_url ? (
-                                            <img
-                                                src={movie.poster_url}
-                                                alt={movie.title ?? 'Poster'}
-                                                loading="lazy"
-                                                className="w-full aspect-[2/3] object-cover"
-                                            />
-                                        ) : null}
-                                        <div className="p-3 flex flex-col">
-                                            <div className="mb-2">
-                                                <h3 className="text-base font-semibold leading-tight">
-                                                    {movie?.title}
-                                                </h3>
-                                                {movie?.director ? (
-                                                    <p className="text-neutral-400 text-xs mt-1">
-                                                        {movie.director}
-                                                    </p>
-                                                ) : null}
-                                                <p className="text-neutral-500 text-xs mt-1">
-                                                    {movie?.duration} min
-                                                </p>
-                                            </div>
-                                            <div className="space-y-1">
-                                                {byDay.map(([_day, dayScreenings]) => (
-                                                    <div
-                                                        key={dayScreenings[0]?.id}
-                                                        className="flex justify-between items-start"
-                                                    >
-                                                        <p className="text-neutral-400 flex-shrink-0 text-xs">
-                                                            {formatDate(dayScreenings[0].start_time)}
-                                                        </p>
-                                                        <div className="flex gap-x-2 flex-wrap justify-end">
-                                                            {dayScreenings
-                                                                .sort(sortScreeningByStartTime)
-                                                                .map((s) => (
-                                                                    <ScreeningTime
-                                                                        key={s.id}
-                                                                        screening={s}
-                                                                    />
-                                                                ))}
-                                                        </div>
-                                                    </div>
-                                                ))}
-                                            </div>
+                                    <FilmPosterCard key={key} movie={movie} className="bg-neutral-700">
+                                        <div className="space-y-1">
+                                            {byDay.map(([_day, dayScreenings]) => (
+                                                <DayRow
+                                                    key={dayScreenings[0]?.id}
+                                                    screenings={dayScreenings}
+                                                />
+                                            ))}
                                         </div>
-                                    </div>
+                                    </FilmPosterCard>
                                 );
                             })}
                         </div>
